Fix SaveSystem and change UI

+++ Metin2D/Assets/Scripts/BossScene/BossController.cs
@@ -14,7 +14,7 @@ public class BossController : MonoBehaviour
     [SerializeField] float waitTime;
     [SerializeField] float backForce;
     [SerializeField] float dashForce;
-    [SerializeField] float health;
+    [SerializeField] int health;
     [SerializeField] float arrowCount;
     [SerializeField] float arrowSpawnTime;
     [SerializeField] float swordJump;
@@ -34,19 +34,19 @@ public class BossController : MonoBehaviour
     bool isBossDie;
     float maxHealth;
     int prevArrow;
-    List<string> attacks = new List<string>() { "jumpAttack", "dashAttack", "shootAttack","swordAttack" };
+    List<string> attacks = new List<string>() { "jumpAttack", "dashAttack", "shootAttack", "swordAttack" };
 
     // Start is called before the first frame update
     void Start()
     {
-        maxHealth=health;
-        powerTime=2;
+        maxHealth = health;
+        powerTime = 2;
         rb = GetComponent<Rigidbody2D>();
         localScaleX = transform.localScale.x;
         direction = (player.transform.position - bossGround.position).normalized;
-        time = Time.time  + 2;
+        time = Time.time + 2;
         healthBar.SetMaxHealth(health);
-        prevRandom=10;
+        prevRandom = 10;
         anim = GetComponent<Animator>();
         swordManScript = transform.Find("sword_man").GetComponent<SwordManAnim>();
         lightBoss = transform.Find("Light 2D").GetComponent<Light2D>();
@@ -84,30 +84,30 @@ public class BossController : MonoBehaviour
                     time = Time.time + 2f;
                     break;
             }
-            
+
             prevRandom = random;
-            
+
         }
         if (health <= 0 && !isBossDie)
         {
             KillBoss();
         }
-        if (health <= maxHealth/2 && powerTime ==2)
+        if (health <= maxHealth / 2 && powerTime == 2)
         {
-            powerTime-=1;
+            powerTime -= 1;
             StopAllCoroutines();
             StartCoroutine(WaitArrow());
             time = Time.time + 9f;
-            isPowerTime=true;
+            isPowerTime = true;
         }
 
-        if (health <= maxHealth/4 && powerTime ==1)
+        if (health <= maxHealth / 4 && powerTime == 1)
         {
-            powerTime-=1;
+            powerTime -= 1;
             StopAllCoroutines();
             StartCoroutine(WaitArrow());
             time = Time.time + 9f;
-            isPowerTime=true;
+            isPowerTime = true;
         }
     }
     void KillBoss()
@@ -115,7 +115,7 @@ public class BossController : MonoBehaviour
         StopAllCoroutines();
         StartCoroutine(LightTimer());
         isBossDie = true;
-        if(rb.bodyType == RigidbodyType2D.Static)
+        if (rb.bodyType == RigidbodyType2D.Static)
         {
             rb.bodyType = RigidbodyType2D.Dynamic;
 
@@ -149,17 +149,20 @@ public class BossController : MonoBehaviour
         SwordAttack();
 
     }
-    IEnumerator WaitJump(){
+    IEnumerator WaitJump()
+    {
         swordManScript.JumpForce();
         yield return new WaitForSeconds(1f);
         JumpAttack();
     }
-    IEnumerator WaitDash(){
+    IEnumerator WaitDash()
+    {
         swordManScript.FrontForce();
         yield return new WaitForSeconds(1f);
         DashAttack();
     }
-    IEnumerator WaitShoot(){
+    IEnumerator WaitShoot()
+    {
         swordManScript.SwordForce();
         yield return new WaitForSeconds(1f);
         ShootAttack();
@@ -168,7 +171,7 @@ public class BossController : MonoBehaviour
     {
         swordManScript.ArrowPowerStart();
         lightBoss.intensity = firstLight;
-        transform.position = new Vector3(-1.3f,6,0);
+        transform.position = new Vector3(-1.3f, 6, 0);
         rb.bodyType = RigidbodyType2D.Static;
         yield return new WaitForSeconds(.6f);
         ArrowAttack();
@@ -177,7 +180,7 @@ public class BossController : MonoBehaviour
     void SwordAttack()
     {
         ChangeDirection();
-        rb.velocity = new Vector2(direction.x * swordJump , -swordJump);
+        rb.velocity = new Vector2(direction.x * swordJump, -swordJump);
         StartCoroutine(ZeroVelocity());
     }
     IEnumerator ZeroVelocity()
@@ -189,7 +192,7 @@ public class BossController : MonoBehaviour
     }
     void JumpAttack()
     {
-      
+
         ChangeDirection();
         rb.AddForce(Vector2.up * jumpForce, ForceMode2D.Impulse);
 
@@ -212,12 +215,12 @@ public class BossController : MonoBehaviour
     }
     IEnumerator EndBackWalk()
     {
-        yield return new WaitForSeconds(waitTime+.5f);
+        yield return new WaitForSeconds(waitTime + .5f);
         swordManScript.EndBackWalk();
     }
     void ShootAttack()
     {
-        
+
         ChangeDirection();
         Instantiate(bullet, shootPoint.position, transform.rotation);
 
@@ -230,9 +233,9 @@ public class BossController : MonoBehaviour
     }
     IEnumerator SetHealthMax()
     {
-        for (int i = 0; i < maxHealth/2; i++)
+        for (int i = 0; i < maxHealth / 2; i++)
         {
-            health +=1;
+            health += 1;
             healthBar.SetHealth(health);
             yield return new WaitForSeconds(.5f);
 
@@ -242,14 +245,14 @@ public class BossController : MonoBehaviour
     {
         for (int i = 0; i < arrowCount; i++)
         {
-            int posX = Random.Range(-15,12);
+            int posX = Random.Range(-15, 12);
             yield return new WaitForSeconds(arrowSpawnTime);
-            while(posX>=prevArrow-1 && posX<=prevArrow+1)
+            while (posX >= prevArrow - 1 && posX <= prevArrow + 1)
             {
-                posX = Random.Range(-15,12);
+                posX = Random.Range(-15, 12);
             }
-            Instantiate(arrow,new Vector2(posX,18),arrow.transform.rotation);
-            prevArrow=posX;
+            Instantiate(arrow, new Vector2(posX, 18), arrow.transform.rotation);
+            prevArrow = posX;
 
         }
         rb.bodyType = RigidbodyType2D.Dynamic;
@@ -258,7 +261,7 @@ public class BossController : MonoBehaviour
     }
     void DashAttack()
     {
-        
+
         ChangeDirection();
         rb.AddForce(Vector2.right * direction * dashForce, ForceMode2D.Impulse);
         StartCoroutine(DAttack());
@@ -291,15 +294,16 @@ public class BossController : MonoBehaviour
             swordManScript.GroundTrue();
 
         }
-        if(other.gameObject.tag =="Wall"){
-            rb.velocity = new Vector2(0,-15f);
+        if (other.gameObject.tag == "Wall")
+        {
+            rb.velocity = new Vector2(0, -15f);
             ChangeDirection();
             swordManScript.EndBackWalk();
         }
     }
     private void OnCollisionExit2D(Collision2D collision)
     {
-        if ( collision.gameObject.tag == "Ground")
+        if (collision.gameObject.tag == "Ground")
         {
             swordManScript.GroundFalse();
         }
