+++ Metin2D/Assets/Scripts/BossScene/HealthBar.cs
@@ -17,16 +17,16 @@ public class HealthBar : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        
+
     }
 
-    public void SetMaxHealth(float health)
+    public void SetMaxHealth(int health)
     {
         slider.maxValue = health;
         slider.value = health;
     }
 
-    public void SetHealth(float health)
+    public void SetHealth(int health)
     {
         slider.value = health;
     }
